Disallow parsing of token with none signature in authenticateToken (#9172)
Signed-off-by: Zixuan Liu <nodeces@gmail.com>

### Motivation

If Apache Pulsar is configured to authenticate clients using tokens based on JSON Web Tokens (JWT), the signature of the token is not validated if the algorithm of the presented token is set to "none". This allows an attacker to connect to Pulsar instances as any user (incl. admins).

### Modifications

- using `parseClaimsJws` instead of `parse`

`parseClaimsJws` can guarantees the correct security model for parsing signed JWTs.

(cherry picked from commit 71bc84164610ff15ff7c5747f923d09b57787e0a)