Fix can't read the latest message of the compacted topic (#14449)
If the reader enabled read compacted and all the data of topic has been compacted
to the compacted ledger, the original topic does not have any data. In this case,
the reader is not able to read the latest message of the compacted topic.

```java
Reader<byte[]> reader = pulsarClient.newReader()
        .topic(topic)
        .startMessageId(MessageId.latest)
        .startMessageIdInclusive()
        .readCompacted(true)
        .create();
```

The root cause is if the `startMessageIdInclusive` is true
and the `startMessageId` is `latest`, the reader will get the
last message ID from the broker and then seek to the last message.
But, the seek method did not consider if there are messages in the
compacted ledger, so not able to seek to last message of the compacted
ledger.

Add force reset option for the managed cursor, if the seek
position < compaction horizon, we should force reset the cursor
to the given position, so that the reader able to start read from
the compacted ledger.

(cherry picked from commit ddc51924e90550ef50fd3cdd099b6aec56aec260)
(cherry picked from commit 956f023cee42dc8471459a80f08858f14939854f)