[Broker] Fix NPE when subscription is already removed (#14363)
* [Broker] Fix NPE when subscription is already removed

* Cover same case for NonPersistentTopic

Master Issue: #14362

There is current a race condition when we remove a subscription. The race and how to reproduce it is described in the #14362. One of the consequences of the race is that there is a chance we try to remove the subscription from the topic twice. This leads to an NPE, as described in the issue.

* Verify that the `sub` is not null before getting its stats.

This is a trivial change.

(cherry picked from commit aee1e7dbc55099c6b7cdc49e7b5e1c4cd66994ce)
(cherry picked from commit af4e000473415c2db3f8406cb78971954f405441)