[fix][broker] Fix creating producer failure when set backlog quota. (#15663)
When trying to reproduce the problem of #15609 using the master's code, it was found that the master also had this bug. The root cause is:
When there is only one ledger in the ManagedLedger, after the current ledger is closed, it has the timestamp and exceeds the time set by the backlog-qutoa, resulting in the failure to create the producer.

The added test could reproduce this.

So when there is only one ledger, we should not exclude it.

If revert this patch, the added test will fail.

(cherry picked from commit 3a8045851f7e9ea62da104dab2b7fe2b47a95ca9)
(cherry picked from commit cc8895e531d872cdba7dcda2f44a9aa1ecd0290b)