Avoid AuthenticationDataSource mutation for subscription name (#16065)
The `authenticationData` field in `ServerCnx` is being mutated to add the `subscription` field that will be passed on to the authorization plugin. The problem is that `authenticationData` is scoped to the whole connection and it should be getting mutated for each consumer that is created on the connection.

The current code leads to a race condition where the subscription name used in the authz plugin is already modified while we're looking at it. Instead, we should create a new object and enforce the final modifier.

(cherry picked from commit e6b12c64b043903eb5ff2dc5186fe8030f157cfc)
(cherry picked from commit 88b51b1e12254ab4feb1e4f0ace5432a330d1eda)