[cleanup][proxy] Remove unused AuthenticationDataSource variable (#19278)
### Motivation

The `AuthenticationDataSource` variable was added to the `ProxyConnection` class in https://github.com/apache/pulsar/pull/1200. It is no longer needed.

Note that we store this variable in the `ServerCnx` because it is used for authorization. Because we do not do authorization in the proxy, we don't need it.

### Modifications

* Remove unused variable

### Verifying this change

This is a trivial change.

### Does this pull request potentially affect one of the following parts:

The only conceivable way this is a breaking change is if a third party implementation implemented their library so that `authState.getAuthDataSource()` has a side effect.

### Documentation

- [x] `doc-not-needed` <!-- Your PR changes do not impact docs -->

No docs needed.

(cherry picked from commit a9b65196258db04ae36bae19b62032acd59a84d8)