[fix][broker] Make ServerCnx#originalAuthData volatile (#19507)
Fixes #19431

`authenticationData` is already `volatile`. We use `originalAuthData` when set, so we should match the style. In #19431, I proposed that we find a way to not use `volatile`. I still think this might be a "better" approach, but it will be a larger change, and since we already use `volatile` for `authenticationData`, I think this is the right change for now.

It's possible that this is not a bug, given that the `originalAuthData` does not change frequently. However, we always want to use up to date values for authorization.

* Add `volatile` keyword to `ServerCnx#originalAuthData`.

This change is a trivial rework / code cleanup without any test coverage.

- [x] `doc-not-needed`

PR in forked repository: skipping test in fork.

(cherry picked from commit c4c174424192202bc0aec9414a1fec0334600fca)
(cherry picked from commit e1d9941656eb0f67c9557b3ae9c1297fbcee7b2d)