Revert "[improve][broker] Authorize originalPrincipal when provided (#19830)"
This reverts commit 0c365e206f4bf12fd47dde649ad7a5fda4f53fdf.

This commit was made as a concession because some were concerned that it
might break implementations. My primary argument was that those
implementations were at risk, so they were okay to "break". Therefore,
I am reverting this on our fork. It is unlikely for there to be many conflicts
because of this diversion.